Migrate search component to TS
