+++ packages/js/components/src/search/autocompleters/index.ts
@@ -14,3 +14,4 @@ export { default as taxes } from './taxes';
 export { default as usernames } from './usernames';
 export { default as variableProduct } from './variable-product';
 export { default as variations } from './variations';
+export * from './types';
