+++ packages/js/components/typings/global.d.ts
@@ -1,0 +1,11 @@
+declare global {
+	interface Window {
+		wcSettings: {
+			variationTitleAttributesSeparator?: string;
+		};
+	}
+}
+
+/*~ If your module exports nothing, you'll need this line. Otherwise, delete it */
+export {};
+
